backend/requirements.txt
fastapi==0.116.1 
uvicorn==0.35.0
pydantic==2.11.7
requests==2.32.5
python-multipart==0.0.20
pillow==11.3.0
cocktaildb

